src/object_receiver_service/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pika>=1.2.0,<1.3.0
minio>=7.1.0,<7.2.0
python-json-logger>=2.0.0,<3.0.0
python-multipart>=0.0.5
python-magic>=0.4.27
libmagic>=0.4.27
